Trying (and failing) to install MonkeyLoader

+++ docker/entrypoint.sh
@@ -1,6 +1,6 @@
 if [ ${STEAM_CRED:-SteamUsername} = "SteamUsername" ] && [ ${STEAM_PASS:-SteamPassword} = "SteamPassword" ]; then
-echo "Required Steam credentials have not been set! Are you sure they are in the \".env\" file? (STEAM_CRED and STEAM_PASS)"
-exit 1
+    echo "Required Steam credentials have not been set! Are you sure they are in the \".env\" file? (STEAM_CRED and STEAM_PASS)"
+    exit 1
 fi
 
 echo "=== INSTALLING HEADLESS CONFIG FILE ==="
@@ -21,10 +21,26 @@ cat >/etc/crystite/conf.d/steamcreds.json <<EOF
 EOF
 /usr/lib/crystite/crystite --install-only --ignore-version-mismatch
 
-if [ ${MODLOADER:=None} != "None" ]; then
+# test -e /var/lib/crystite/Resonite/modloader && LAST_MODLOADER=$(cat /var/lib/crystite/Resonite/modloader)
+# if [ ${MODLOADER:=None} != ${LAST_MODLOADER:=None} ]; then
+# echo "=== CLEANING UP PREVIOUS MODLOADER ==="
+# case $LAST_MODLOADER in
+#   RML)
+#     test -e /etc/crystite/conf.d/modloader.json && rm /etc/crystite/conf.d/modloader.json
+#     # This probably isn't necessarry but lets do it anyway
+#     rm /var/lib/crystite/Resonite/Libraries/ResoniteModLoader.dll /var/lib/crystite/Resonite/rml_libs/0Harmony.dll
+#     ;;
+#   MonkeyLoader|MonkeyLoaderRML)
+    
+#     ;;
+# esac
+# fi
+# echo $MODLOADER > /var/lib/crystite/Resonite/modloader
+
+if [ $MODLOADER != "None" ]; then
 echo "=== INSTALLING MODLOADER ==="
 case $MODLOADER in
-  RML)
+  ResoniteModLoader|RML)
     echo "Selected modloader: ResoniteModLoader"
     mkdir \
         /var/lib/crystite/Resonite/Libraries \
@@ -32,20 +48,38 @@ case $MODLOADER in
         /var/lib/crystite/Resonite/rml_libs \
         /var/lib/crystite/Resonite/rml_config
     wget -O /var/lib/crystite/Resonite/Libraries/ResoniteModLoader.dll "https://github.com/resonite-modding-group/ResoniteModLoader/releases/latest/download/ResoniteModLoader.dll"
+    test ! -e /var/lib/crystite/Resonite/Libraries/ResoniteModLoader.dll && echo "Failed to download ResoniteModLoader.dll!" && exit 1
     wget -O /var/lib/crystite/Resonite/rml_libs/0Harmony.dll "https://github.com/resonite-modding-group/ResoniteModLoader/releases/latest/download/0Harmony.dll"
+    test ! -e /var/lib/crystite/Resonite/rml_libs/0Harmony.dll && echo "Failed to download 0Harmony.dll!" && exit 1
     cat >/etc/crystite/conf.d/modloader.json <<EOF
 {
     "Resonite": {
-        "pluginAssemblies": "/var/lib/crystite/Resonite/Libraries/ResoniteModLoader.dll",
+        "pluginAssemblies": ["/var/lib/crystite/Resonite/Libraries/ResoniteModLoader.dll"],
     },
 }
 EOF
     ;;
   MonkeyLoader)
     echo "Selected modloader: MonkeyLoader"
+    echo "Not implemented" && exit 1
+    ML_URL=$(wget -O- https://api.github.com/repos/ResoniteModdingGroup/MonkeyLoader.GamePacks.Resonite/releases/latest | jq -r '.assets[] | select(.name|endswith(".zip")) | .browser_download_url')
+    test ! -v ML_URL && echo "Failed to get MonkeyLoader download URL!" && exit 1
+    wget -O /var/lib/crystite/MonkeyLoader.zip "$ML_URL"
+    test ! -e /var/lib/crystite/MonkeyLoader.zip && echo "Failed to download MonkeyLoader.zip!" && exit 1
+    unzip -o /var/lib/crystite/MonkeyLoader.zip -d /var/lib/crystite/Resonite
+    rm /var/lib/crystite/MonkeyLoader.zip /var/lib/crystite/Resonite/MonkeyLoader/GamePacks/MonkeyLoader.GamePacks.ResoniteModLoader.nupkg
     ;;
   MonkeyLoaderRML)
     echo "Selected modloader: MonkeyLoader + ResoniteModLoader"
+    echo "Not implemented" && exit 1
+    ML_URL=$(wget -O- https://api.github.com/repos/ResoniteModdingGroup/MonkeyLoader.GamePacks.Resonite/releases/latest | jq -r '.assets[] | select(.name|endswith(".zip")) | .browser_download_url')
+    test ! -v ML_URL && echo "Failed to get MonkeyLoader download URL!" && exit 1
+    wget -O /var/lib/crystite/MonkeyLoader.zip "$ML_URL"
+    test ! -e /var/lib/crystite/MonkeyLoader.zip && echo "Failed to download MonkeyLoader.zip!" && exit 1
+    unzip -o /var/lib/crystite/MonkeyLoader.zip -d /var/lib/crystite/Resonite
+    rm /var/lib/crystite/MonkeyLoader.zip
+    # rm /var/lib/crystite/MonkeyLoader.zip /var/lib/crystite/Resonite/MonkeyLoader/GamePacks/MonkeyLoader.GamePacks.ResoniteModLoader
+    # wget -O /var/lib/crystite/Resonite/MonkeyLoader/GamePacks/MonkeyLoader.GamePacks.ResoniteModLoader "https://github.com/ResoniteModdingGroup/MonkeyLoader.GamePacks.ResoniteModLoader/releases/latest/download/MonkeyLoader.GamePacks.ResoniteModLoader.nupkg"
     ;;
   *)
     echo "Unrecognized modloader \"$MODLOADER\", skipping installation."
